requirements.txt
pytest==7.4.0
playwright==1.39.0
pytest-playwright==0.4.0
allure-pytest==2.13.2
pydantic==2.0.3
python-dotenv==1.0.0

